requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
numpy==1.21.5
pandas==1.4.2
liac-arff==2.5.0




